requirements.txt
numpy==2.2.2
pillow==11.1.0
pydantic==2.10.6
PyYAML==6.0.2
scikit-learn==1.6.1
torch>=2.7.0
torchvision>=0.22.0
tqdm==4.67.1
mypy==1.15.0
pytest==8.3.4
pytest-cov==6.0.0
ruff==0.9.4
bandit==1.8.2
pip-audit==2.8.0
colorama==0.4.6

# Optional for future backend parity (scaffold-only in v1):
# tensorflow==2.18.0
